refactor loader and text_splitter package inside preprocess package
